Leave one problem for importing file modules.
- Also fixed or (temporarily) suppressed some warnings.
- Support specifying external module paths with `use some_module in "path.to.module.pest"` in grammar.
- Fixed conversions from module paths to file system paths.
- Refactor some codes.
- Remove dev-dependency `lazy-static` for pest3_generator.

A problem that I meet is that generated enum `Rule` from imported modules may not be used in modules that they are imported into.
Say, there is two modules `a` and `b`, and a is importing `b`. Maybe I need to add a variant `b(b::Rule)` in `enum Rule` generated for `a`, like

```
enum Rule {
    b(rules::b::Rule),
    ...
}
pub mod rules {
    pub mod b {
        pub enum Rule {
            ...
        }
    }
    ...
}
```

However, definitions of some traits need also to be changed. I may need some time to evaluate this. @tomtau What's your point?